refactor: storeInfo.js
