+++ src/const/support.js
@@ -1,0 +1,28 @@
+// 직영점
+export const MAP_CLIENT_ID = process.env.REACT_APP_MAP_CLIENT_ID || 'i1nv00h8zo';
+export const COORDINATE = {
+  lat: 37.52369,
+  lng: 127.03901
+};
+export const storeImages = [
+  {
+    src:'../../images/event/img_store1.jpg',
+    alt: '소니스토어 압구정점 내부 이미지 1'
+  },
+  {
+    src:'../../images/event/img_store2.jpg',
+    alt: '소니스토어 압구정점 내부 이미지 2'
+  },
+  {
+    src:'../../images/event/img_store3.jpg',
+    alt: '소니스토어 압구정점 내부 이미지 3'
+  },
+  {
+    src:'../../images/event/img_store4.jpg',
+    alt: '소니스토어 압구정점 내부 이미지 4'
+  },
+  {
+    src:'../../images/event/img_store5.jpg',
+    alt: '소니스토어 압구정점 내부 이미지 5'
+  },
+];
